+++ src/CoreLogic/ExprCalc.CoreLogic/CoreLogicRegistrationExtensions.cs
@@ -6,6 +6,7 @@ using ExprCalc.CoreLogic.Resources.CalculationsRegistry;
 using ExprCalc.CoreLogic.Resources.ExpressionCalculation;
 using ExprCalc.CoreLogic.Services.CalculationsProcessor;
 using ExprCalc.CoreLogic.Services.RegistryRepopulation;
+using ExprCalc.CoreLogic.Services.StorageCleanup;
 using ExprCalc.CoreLogic.UseCases;
 using ExprCalc.Storage.Api.Repositories;
 using Microsoft.Extensions.DependencyInjection;
@@ -33,6 +34,7 @@ namespace ExprCalc.CoreLogic
 
             serviceCollection.AddHostedService<CalculationsProcessingService>();
             serviceCollection.AddHostedService<RegistryRepopulationJob>();
+            serviceCollection.AddHostedService<StorageCleanupService>();
         }
 
         public static void AddCoreLogicMetrics(this MetricsRegistry registry)