+++ src/Storage/ExprCalc.Storage.Api/Repositories/ICalculationRepository.cs
@@ -18,6 +18,7 @@ namespace ExprCalc.Storage.Api.Repositories
         Task<bool> ContainsCalculationAsync(Guid id, CancellationToken token);
         Task<Calculation> AddCalculationAsync(Calculation calculation, CancellationToken token);
         Task UpdateCalculationStatusAsync(CalculationStatusUpdate calculationStatusUpdate, CancellationToken token);
-        Task<int> ResetNonFinalStateToPending(DateTime maxCreatedAt, DateTime newUpdatedAt, CancellationToken token);
+        Task<int> ResetNonFinalStateToPendingAsync(DateTime maxCreatedAt, DateTime newUpdatedAt, CancellationToken token);
+        Task<int> DeleteCalculationsAsync(DateTime createdBefore, CancellationToken token);
     }
 }